Created anchor button component and replaced DownloadLink
