+++ webapp/client/src/components/AnchorButton.test.js
@@ -1,15 +1,15 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import ButtonLink from "./ButtonLink";
+import AnchorButton from "./AnchorButton";
 
 const MOCK_PROPS = {
   text: "anchor text",
   url: "/some/link/path",
 };
 
-describe("ButtonLink", () => {
+describe("AnchorButton", () => {
   it("should render the text", () => {
-    render(<ButtonLink {...MOCK_PROPS} />);
+    render(<AnchorButton {...MOCK_PROPS} />);
 
     expect(screen.getByText(MOCK_PROPS.text)).toBeInTheDocument();
     expect(screen.getByText(MOCK_PROPS.text).closest("a")).toHaveAttribute(
